move controller back to runner-ct, properly export and consume
