--- packages/runner-ct/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/runner-ct')
